Add tests for run_scraper.
