+++ test_run_scraper.py
@@ -1,0 +1,68 @@
+#!/usr/bin/env python
+# Copyright 2017 Scraper Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# No docstrings required for tests, and tests need to be methods of classes to
+# aid in organization of tests. Using the 'self' variable is not required.
+#
+# pylint: disable=missing-docstring, no-self-use, too-many-public-methods
+
+"""Tests for run_scraper, representing end-to-end tests of scraper as a whole."""
+
+import unittest
+
+import freezegun
+import mock
+import testfixtures
+
+import run_scraper
+
+
+class TestRunScraper(unittest.TestCase):
+
+    def test_args(self):
+        rsync_host = 'mlab1.dne0t.measurement-lab.org'
+        rsync_module = 'ndt'
+        data_dir = '/tmp/bigplaceforbackup'
+        rsync_binary = '/usr/bin/rsync'
+        rsync_port = 1234
+        max_uncompressed_size = 1024
+        args = run_scraper.parse_cmdline([
+            '--rsync_host', rsync_host, '--rsync_module', rsync_module,
+            '--data_dir', data_dir, '--rsync_binary', rsync_binary,
+            '--rsync_port', str(rsync_port), '--max_uncompressed_size',
+            str(max_uncompressed_size)
+        ])
+        self.assertEqual(args.rsync_host, rsync_host)
+        self.assertEqual(args.rsync_module, rsync_module)
+        self.assertEqual(args.data_dir, data_dir)
+        self.assertEqual(args.rsync_binary, rsync_binary)
+        self.assertEqual(args.rsync_port, rsync_port)
+        self.assertEqual(args.max_uncompressed_size, max_uncompressed_size)
+        args = run_scraper.parse_cmdline([
+            '--rsync_host', rsync_host, '--rsync_module', rsync_module,
+            '--data_dir', data_dir
+        ])
+        self.assertEqual(args.rsync_binary, '/usr/bin/rsync')
+        self.assertEqual(args.rsync_port, 7999)
+        self.assertEqual(args.max_uncompressed_size, 1000000000)
+
+    def test_args_help(self):
+        with self.assertRaises(SystemExit):
+            with testfixtures.OutputCapture() as _:
+                run_scraper.parse_cmdline(['-h'])
+
+
+if __name__ == '__main__':  # pragma: no cover
+    unittest.main()
